requirements.txt
pydantic>=1.8.1
Flask>=1.1.2
pytest>=6.2.3
requests>=2.25.1
psycopg2-binary>=2.8.6